+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Changelog for package rosros
 0.2.0 (2022-12-09)
 -------------------
 * support dictionaries in publish and service calls
+* return node object from rosros.init_node() in ROS2
 
 0.1.0 (2022-10-30)
 -------------------